requirements.txt
asgiref==3.8.1
Django==5.1.4
django-jazzmin==3.0.1
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.3
